Created parameter for canvas resolution
It is possible now to set the canvas resolution in the editor.
The texture size will be derived from this resolution, respecting the
aspect ratio of the canvas mesh / actor.
The resolution acts as if it is in world space: same resolution values
for canvases with different shapes or sizes will produce the same
quality on screen.