Added player colors
Each player has an ID (as of now, a player spawns with ID = 0, can edit
it on the editor)
Each ID has an associated color.
Ink drops will paint the canvas with the color associated to the ID of
the player that spawned the ink drops.
TODO: Give ink drop particles the color associated to their ID (right
now they're always red, although they do paint the canvas with the
correct color).